Fixed detection of unsigned add overflow
