+++ src/test/java/org/simdjson/NumberParsingTest.java
@@ -17,7 +17,7 @@ import java.util.stream.Stream;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.simdjson.JsonValueAssert.assertThat;
-import static org.simdjson.StringUtils.toUtf8;
+import static org.simdjson.TestUtils.toUtf8;
 
 public class NumberParsingTest {
