--- src/test/java/org/simdjson/StringUtils.java
@@ -1,26 +1,0 @@
-package org.simdjson;
-
-import jdk.incubator.vector.ByteVector;
-
-import java.util.Arrays;
-
-import static java.nio.charset.StandardCharsets.UTF_8;
-
-class StringUtils {
-
-    static String padWithSpaces(String str) {
-        byte[] strBytes = toUtf8(str);
-        byte[] padded = new byte[strBytes.length + 64];
-        Arrays.fill(padded, (byte) ' ');
-        System.arraycopy(strBytes, 0, padded, 0, strBytes.length);
-        return new String(padded, UTF_8);
-    }
-
-    static ByteVector chunk(String str, int n) {
-        return ByteVector.fromArray(StructuralIndexer.SPECIES, str.getBytes(UTF_8), n * StructuralIndexer.SPECIES.vectorByteSize());
-    }
-
-    static byte[] toUtf8(String str) {
-        return str.getBytes(UTF_8);
-    }
-}
